feat: Docker 연결 설정 및 JPA 리포지토리 구조 개선
- Docker TCP 연결을 위한 설정 변경
- JPA 리포지토리와 QueryDSL 구현체 분리
- Container 리포지토리 중복 제거
- file_path nullable 설정

+++ backend/src/main/java/solid/backend/Directory/dto/DirectoryDelDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class DirectoryDelDto {
     private Integer directoryId;
-    private Integer containerId;
+    private Long containerId;
     private String directoryRoot;
     private String directoryName;
 }
